requirements.txt
streamlit
joblib
pandas
python-dotenv
google-generativeai